requirements.txt
importlib_metadata;python_version<"3.10"
importlib_resources;python_version<"3.10"
PyYAML
